feat(View): 겜블러 이름과 히트/스테이 여부 입력받는 기능 구현
