--- src/main/java/blackjack/blackjack.view/InputView.java
@@ -1,20 +1,0 @@
-package blackjack.blackjack.view;
-
-import static java.util.stream.Collectors.toList;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.Scanner;
-
-public class InputView {
-    private static final Scanner SCANNER = new Scanner(System.in);
-
-    public List<String> scanPlayerNames() {
-        System.out.println("게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)");
-
-        return Arrays.stream(SCANNER.nextLine()
-            .split(",", -1))
-            .map(String::trim)
-            .collect(toList());
-    }
-}
